refactor: 설정파일 파일 구조 변경
